requirements.txt
flask
pandas
openpyxl
google-generativeai
flask-cors
dotenv